Switched Framer Generator to Framer Classic
Could not immediately understand how to get Framer Generator to work, but found that Framer Classic import method and saving seemed to do the job. This comes with payment or a strict 14 day trial, but does at least successfully bring sketch files into the necessary format to open in Unity with the plugin. More work is needed to determine if Framer Generator can work on its own this way.