Added import scale, fixed for 2018.1, and updated readme (#2)
* Switched Framer Generator to Framer Classic

Could not immediately understand how to get Framer Generator to work, but found that Framer Classic import method and saving seemed to do the job. This comes with payment or a strict 14 day trial, but does at least successfully bring sketch files into the necessary format to open in Unity with the plugin. More work is needed to determine if Framer Generator can work on its own this way.

* Readme re-revised to include both instructions

Framer Generator and Framer Classic now both supported in readme.

* Import scale functionality added, and updated to run in 2018.1

* quick typo fix to formatting